Add a new metric function 'observeBucketed', for generating counter values tagged according to a set of buckets.
This simulates the behavior of the prometheus histogram metric, and
will generate tags for observed values according to a set of predefined
buckets.

For example, the call:
        'observeBucketed('foo.bar', 45, [10, 25, 50, 100])'

will generate a counter metric with the tags ['le:50', 'le:100'],
since it is less than those values, along with a special 'le:Inf'
tag that all values receive (even if they exceed the largest
configured bucket). Since the value 45 is greater than the
bucket '10', an 'le:10' tag will not be present.

This change also includes an update to the sinon library, and
fixes some tests that broke as a result of a change in the
way Error instances are compared.